pyproject.toml
[project]
name = "repl-nix-slidegenai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.3",
    "pandas>=2.2.3",
    "python-pptx>=1.0.2",
    "streamlit>=1.40.0",
]
